Updated the settings views to use Swift and storyboards (#118)
Storyboards

The initial storyboard is the main settings view. This is a static tableview.
The second storyboard is the sync settings view. This is a dynamic tableview where we can have various syncing services with different settings associated to each one.
Swift

Code is pretty much the same as obj-c but just change to Swift syntax.